Refactor top level of GoAT library; Improve diagrams
